LayoutTests/inspector/audit/audit-test-suite.html

<!doctype html>
<html>
<head>
<script src="../../http/tests/inspector/resources/inspector-test.js"></script>
<script src="./resources/audit-test-fixtures.js"></script>
<script>
function test()
{

    let suite = InspectorTest.createAsyncSuite("AuditTestSuite");

    let auditTestSuite1 = new testSuiteFixture1("FakeTestSuite", "FakeTestSuite");
    let auditTestSuite2 = new testSuiteFixture1("FakeTestSuite", "FakeTestSuite");

    suite.addTestCase({
        name: "AuditTestSuite Id",
        description: "Should exist and be unique",
        async test() {
            InspectorTest.expectEqual(typeof(auditTestSuite1.id), "symbol", "AuditTestSuite1 has ID with correct type.");
            InspectorTest.expectEqual(typeof(auditTestSuite2.id), "symbol", "AuditTestSuite2 has ID with correct type.");
            InspectorTest.expectThat(auditTestSuite1.id !== auditTestSuite2.id, "AuditTestSuites with same name have different unique IDs.");
        }
    });

    suite.addTestCase({
        name: "AuditTestSuite testCaseCount",
        description: "Should represents correct number of test case.",
        async test() {
            InspectorTest.expectEqual(auditTestSuite1.testCases.length, 2, "There are two tests.");
        }
    });



    suite.addTestCase({
        name: "AuditTestSuite should run tests sequentially.",
        description: "Tests should be ran in the order that was defined in the testDescriptor.",
        async test() {
                
            let testOrderSymbol = Symbol("test-order");
            window[testOrderSymbol] = 0;

            let TestOrderSuite = class TestOrderSuite extends WI.AuditTestSuite 
            {
                static testCaseDescriptors()
                {
                    return [
                        {
                            name: "test 1",
                            description: "order 1",
                            async test() {
                                InspectorTest.expectEqual(window[testOrderSymbol], 0, "First test is ran.");
                                window[testOrderSymbol] = 1;
                            }
                        },
                        {
                            name: "test 2",
                            description: "order 2",
                            async test() {
                                InspectorTest.expectEqual(window[testOrderSymbol], 1, "Second test is ran.");
                                window[testOrderSymbol] = 2;
                            }
                        },
                        {
                            name: "test 3",
                            description: "order 3",
                            async test() {
                                InspectorTest.expectEqual(window[testOrderSymbol], 2, "Third test is ran.");
                                window[testOrderSymbol] = 3;
                            }
                        },
                        {
                            name: "test 4",
                            description: "order 4",
                            async test() {
                                InspectorTest.expectEqual(window[testOrderSymbol], 3, "Fourth test is ran.");
                                window[testOrderSymbol] = 4;
                            }
                        },
                        {
                            name: "test 4",
                            description: "order 4",
                            async test() {
                                InspectorTest.expectEqual(window[testOrderSymbol], 4, "Last test is ran.");
                            }
                        }
                    ]
                }
            }

            let auditManager = new WI.AuditManager;
            auditManager.addTestSuite(TestOrderSuite);

            let result = auditManager.runAuditTestByRepresentedObject(auditManager.testSuites[0]);
 
        }
    });

    suite.runTestCasesAndFinish();
}
</script>
</head>
<body onload="runTest()">
<p>Test for the AuditTestSuite.</p>
</body>
</html>
