LayoutTests/inspector/audit/audit-test-suite-expected.txt

Test for the AuditTestSuite.


== Running test suite: AuditTestSuite
-- Running test case: AuditTestSuite Id
PASS: AuditTestSuite1 has ID with correct type.
PASS: AuditTestSuite2 has ID with correct type.
PASS: AuditTestSuites with same name have different unique IDs.

-- Running test case: AuditTestSuite testCaseCount
PASS: There are two tests.

-- Running test case: AuditTestSuite should run tests sequentially.
PASS: First test is ran.
PASS: Second test is ran.
PASS: Third test is ran.
PASS: Fourth test is ran.
PASS: Last test is ran.
